requirements.txt
Django~=4.0.6
requests~=2.28.1
djangorestframework~=3.13.1
pytest-django==4.5.2
psycopg2-binary==2.9.3
psycopg2==2.8.6

